// packages/mrbd/src/mrbd/ui/InputTextViewLayout.types.ts
/**
 * Copyright (c) Meta Platforms, Inc. and affiliates.
 *
 * This source code is licensed under the Apache License, Version 2.0 found in the
 * LICENSE file in the root directory of this source tree.
 *
 */

import type { CSSProperties } from 'react';

export interface InputTextViewLayoutState {
  hasText: boolean;
  showAccessory: boolean;
  showActionButton: boolean;
  containerWidth: CSSProperties['width'];
  editAreaStyle: CSSProperties;
}

export interface InputTextViewScrollState {
  isScrollable: boolean;
  showTopFade: boolean;
  showBottomFade: boolean;
  scrollRatio: number;
  viewportRatio: number;
}